move ToastContainer to separate file
