+++ packages/components/src/toast/index.tsx
@@ -1,1 +1,3 @@
-export { Toast, ToastContainer, useToast } from './toast'
+export type { ToastProps } from './toast'
+export { Toast } from './toast'
+export { ToastContainer, useToast } from './toast-container'
